+++ src/main/java/com/github/jlangch/venice/util/ipc/IMessage.java
@@ -44,6 +44,8 @@ import com.github.jlangch.venice.impl.types.VncVal;
  * ├───────────────────────────────┤
  * │ ExpiresAt                     │   client (may be null)
  * ├───────────────────────────────┤
+ * │ Timeout                       │   client (server-side queue offer/poll timeout)
+ * ├───────────────────────────────┤
  * │ Request ID                    │   client (may be used for idempotency checks by the receiver)
  * ├───────────────────────────────┤
  * │ Topic                         │   client