+++ ChangeLog.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - Added :openai module support for assistant run api (...)
 - Added :openai module support for assistant run steps api (...)
 
+- Improved the script auto-run Venice JAR rewriter